Added more informative error messages when we get a weird return from NSG.  NSG was failing in odd ways, which I _finally_ sussed was a result of using the wrong d1mach.f file, which was causing NSG to go haywire and give impossible return codes.  This required a change to the makefile to manage the correct one.  I've verified the changes by building export_source and rebuilding from the result.
d1mach.*: d1mach.f removed from CVS, as it is a derived file and not a
source one; d1mach.f_* added: d1mach.f_master is the one from which
the other two are derived.  None of these should need maintenance.

gaufit2.f modified so that the various error return codes from NSG
have individual err_rep messages.  When NSG returns with one of the
impossible return codes, at least gaufit reports it properly.

makefile modified so that d1mach.f is managed as a link to
d1mach.f_$(SOURCE_VARIANT).  The d1mach.f_*ieee are now the ones
packaged up.

mk modified so that the TAR_IN variable does not include the `v'
option (because that's just silly).

svn path=/trunk/; revision=5195