Mostly changes relating to documentation.  Made headers of bin/sgml2... conform to programcode DTD, and added sgmlbin.sgml as a shell, an instance of the programcode DTD, which includes SGML fragments generated from the scripts.  The lib/Makefile main target keeps this documentation up-to-date.
Added doc/documents/sun124.summary, and added it to the CATALOG.

In dtd2sgml, distinguished elemref by adding elemxref for referring to
elements in another DTD.

Assorted changes to ssn70.sgml, to bring in the newly-generated lib
and bin documentation.  Made the actions which generate this invoke
the makefiles in bin and lib.

Supported moduletype in HTML D-C.

Supported moduletype in LaTeX D-C.

svn path=/trunk/; revision=6270