A small change to process-backmatter (adding a guard against being called in the context of the document element).
In make-idindex, I now have it selecting elements in
(idindex-element-list) rather than (target-element-list).  This
function seems to be taking an atrociously long time to run (294s
runtime with `--jadeflags -Vsuppress-idindex' and 759s without), which
I just don't understand -- what can be happening in there?  I thought
it might be due to taking a long time to find <routine> elements
cross-document, so had idindex-element-list consist of everything in
target-element-list except "routine", but that didn't make any
difference, so now the two are identical.  Tracing the output (by
putting (debug) statements in the handlers in make-idindex-mode) suggests
that every element in the idindex is taking a couple of seconds or so
to process.  Bizarre!

svn path=/trunk/; revision=8052