Changed t1.correct, so that the orientation and perpendicularity in the decomposition of the linear fit have the correct signs (also recognised adjustment in output format of times).  These signs were wrong before because there was a bug in sla_dcmpf which I confirmed with Pat in December 2001.  I've verified that t1.correct is _now_ correct, since I've recalculated pltcon(1..6,nsol) from the decomposed values (around line 2206 in a scratch edit of astrml.f) and confirmed that we get back the numbers we put in.
t2.c is a collection of the test cases which threw up this bug, as
tested against the fortran library version of sla_dcmpf_, and against
an internal version of slaDcmpf.  If this test fails, it probably
means that the slalib on that machine is out-of-date.  The slaDcmpf
embedded in t2.c is that which Pat mailed me on 2001 December 19 (to
build this version, you need to have slaDrange from a slalib_c
distribution).

svn path=/trunk/; revision=10730