Fixed problems with stdargs va_list usage.
Although a va_list can be passed as an argument into a function, that function can only use the va_args function on the va_list, you cannot use va_start & va_end to preserve the given reference, or directly make a copy of the reference (the type of va_list is hidden, so what a copy of the reference means
is undefined, but later compilers could make use of va_copy to do this).

This meant that the repeated use of va_list in Et_VaEval had be unrolled
and a different function introduced to estimate the expanded Tcl command-line
length.

Removed all the K & R specific code. This was clearly broken anyway.

svn path=/trunk/; revision=15083