Add workaround for NFS file mapping problem under Linux. Without this the mapped data is not flushed back to disk.
General ramblings about this problem.

This issue may come down to how the the UNIX specification says
syncronization of data is handled. That is how to make sure it is
written to disk or not, when mapped and then modified using MAP_SHARED.
Opinions seem to vary in the interpretation.

If the strong opinion is correct that you must msync (or fsync) to ensure
correct data synchronization, then I think an msync should be
done just before the call to munmap in rec1_unmap_file.c, rather than
a full sync here, but asynchronous (MS_ASYNC) msyncs do not work with
the current problem (seem in Linux kernels 2.6.9, 2.6.10 and 2.6.11),
so there is a potentially large performance hit (if we used MS_SYNC).
Better to sync just once for now.

svn path=/trunk/; revision=16910