Change how the -frameset option is handled. This can now have more than one input and output axes, so that a suitable lutmap can be generated for plots that do not invert.
Add -mapping to allow access to the 1D spectral transformation (used
instead of -frameset).

Add -axis to specify which axis of the frameset is the spectral one.

Change the plot transformation so that it connects via the base frame
not the current one. This gives a linear transformation from pixel to
graphics (and matches the behaviour of SPLAT & LINPLOT). Plots which
are not linear in physical coordinates will now look non-linear.

svn path=/trunk/; revision=23155