prm: changes to make clear typeless BOZ usage
Backout of some of the latest changes to make-prm-par. We do not want
to use Z, that is usually a typed BOZ so we may as well write plain
numbers. Make the various descriptions of all this consistent.

Move the STAR_CNF_COMPATIBLE_SYMBOLS check to after the BOZ checks.
In fact that was always true (since the BOZ checks are run as part of
STAR_CNF_COMPATIBLE_SYMBOLS), but appeared otherwise.