gaia: remove bootstrap of TEA directories
Background: after buildsystem changes, like version numbers, it is neceessary to regenerate gaia/ and gaia3d/, this was done by the normal bootstrap, but that idea is flawed as it means a re-bootstrapped GAIA will create modified configure scripts with the current prefix burnt in, which are not changes we want to see. Back to regenerating by hand.