specx: Tweak build system for ifort/icc
Specx could not be linked with ifort due to problems with multiple
main definitions. The intention was to link with the C compiler
but that was not happening because the specx_SOURCES definition included
both C and fortran source files. Rejigged slightly so that the fortran
source file is compiled into a convenience library. This then allows
the system to link with the C compiler.