smf_calcmodel_com: Don't use common-mode to modify flatfield
Rather than fitting the relative gains in each detector to given
the same amplitude of the common-mode signal, scale the common
mode to the data. Seemed like a good idea before, but it was
executed incorrectly. Once the flatfield task works properly
(with calibrated heater resistances) this will be unnecessary
anyways.