smurf: Iterative map-maker can evaluate models after AST (map)
The MODELORDER keyword must now explicitly give AST as one of
the components (which estimates the map and its time-domain
projection). It is possible to specify model components
after AST. This is handy for removing ringing around bright
sources: just do the filtering (FLT) after the map (AST). The
example dimmconfig has been updated to do this. Also, FLT
and DKS seem to be degenerate and chi^2 blows-up, so DKS is
currently turned off.