smurf: move stationary data flagging to smf_model_create
Since the creation of .DIMM files for the memiter=0 case of the
iterative map-maker does not propagate the JCMTState header
information, previously smf_iteratemap was unable to flag
stationary data. By passing 'flagstat' to smf_model_create
flagging of stationary pointing can be accomplished in
a uniform way for both the memiter=0 and memiter=1 cases.