smurf: optionally use multiple threads in smf_filter_execute
If a workforce pointer is supplied to smf_filter execute it
will now filter blocks of bolos in parallel. However, you
can still call it with a NULL workforce pointer and it
will run in a single thread (it is thread safe so it can
be called from a higher-level parallelized routine).
Currently getting sporadic segv's with smurf_sc2threadtest
on test 4 which attempts to supply a workforce poiner.