init: when relocating perl site_perl should come second
Previously we add the path to the relocated library perl path
to PERL5LIB and then prepended site_perl. This had the unforseen
side effect of adding site_perl in front of standard perl and
broke the lehuakona build that had Errno.pm from 5.8.8 in the
site_perl tree being picked up in front of the Errno.pm from 5.10.0
from the standard perl tree. The 5.8.8 version was not compatible
with the perl binary since the kernel had changed.