atl: atl_mgfts was removing non-blank cards by mistake
At the end of atl_mgfts it attempts to remove contiguous
blank lines. This was not working properly because AST_FINDFITS
was incrementing the Card to the following card and so AST_DELFITS
was removing the card following two contiguous blank cards.

Fixed to decrement the Card attribute prior to call AST_DELFITS.