ccdpack: Remove local kaplibs routines and use KAPLIBS instead
Most of the kaplibs routines were more or less the same as
the kaplibs versions. Bug fixes have been transferred to
kaplibs independently.

One note is that KPG1_ISCLx in kaplibs takes a <T> UPPER
and LOWER values rather than DOUBLE. The one CCDPACK
routine using the _DOUBLE argument has been modified.
The local CCDPACK patch was probably best (since internally
the argument is immediately cast to a double) but the
extensive usage in KAPPA means that change is deferred.