kaplibs: Move ccd1_nxwrd out of ccdpack and polpack into kaplibs
CCD1_NXWRD is a wrapper around CHR_FIW[ES] and could be generically
useful in many places (including kappa). There was a clone in polpack
that was using it so it seems a good way to remove a duplicate is
to move it into kaplibs now and then proceed to replace CHR_FIW[ES]
usage in the rest of the software.