Change location of perl git repository
I have cloned the perl git repository and placed it in the same
location as the rest of the Starlink repositories. This allows
a relative URL to be used so that it will work with both git:
and http: protocols matching the rest of the checkouts rather
than requiring that both git and http protocols will work.