kaplibs: Fix loop termination condition in kpg1_ptnd
The PIX and CUBND arrays are checked in a WHILE loop. It is
possible for the index in that array check to exceed the
number of dimensions and address uninitialised memory. This
patch applies the same test to prevent buffer overrun to
the general case of J exceeding NDIM. Valgrind no longer
warns.