KAPLIBS: Do not count collapsed data comprising only bad values.
The count of the output array's bad elements generated by the MINPIX
threshold should exclude those not attributable to the threshold.
This is where all the contributing input elements used to derive an
output element are bad.

This is in line with the original intention and as documented in
KAPPA:COLLAPSE.