mers: Add msgIfgetenv for reading message reporting level from environment
Split msgIfget_adam into a routine that retrieves the string from the ADAM
parameter and a routine that converts a string into a message reporting
level. The new routine then calls getenv() to get the environment variable
value and passes it to the msg1Ifget internal routine for conversion
and setting of the message reporting level.