KAPPA: Remove QUIET parameter from BEAMFIT.
Should now set MSG_FILTER to QUIET to suppres display of results.

This is the first of about a dozen tasks requiring this modification.
Dealing with this one separately to resolve merge conflict, after
Tim's (as it turns out) unnecessary revision for MSG_IFLEV's API,
while I was waiting for MSG_FLEVOK.