agi: Allow for longer string for AGI_USER file name
64 characters is too short for AGI_USER/AGI_NODE.sdf on some
systems that auto-generate temporary directory locations to
be unpredictable. Modify all routines using file names to
read a constant from AGI1_PAR and make that constant large.

Also modify agi1_fname.f to use CHR_LEN to find the length
of the string rather than searching for a space. The latter
does not work when the string fills the buffer.