mers: msgIfget now falls back to environment if parameter retrieval fails
If the parameter can not be obtained (or if it is set to PAR__NULL)
the environment variable MSG_FILTER will now be read via msgIfgetenv.

This allows applications that currently use parameters to still use
them explicitly but also fall back to the value in use for applications
such as KAPPA for consistency.