pcs: Trick IFL parser to believe that MSG_FILTER and QUIET parameters exist
Modify the IFL parser so that at the end of every interface definition
two new parameters are added. This allows MSG_IFGET in the fixed
part to always have access to MSG_FILTER and QUIET parameters
to handle message filtering without forcing every application
to have private definitions in IFL files.

Does not overwrite definitions for QUIET and MSG_FILTER
parameters that do exist in the IFL files.

This patch requires that all IFC files are recompiled.