KAPPA: Correct typo's in SUN/95 Figure 12.
The final grid co-ordinate was erroneously set to 6.5 instead of 5.5.
Since the original Publisher file is not available, the encapsulated
PostScript has been editied with GIMP.  The typeface is not a perfect
match but is adequate for the purpose.  Also inserted a missing hyphen.

There was a side effect of the image manipulation.  The dark shadow
effect was replaced by a patterned edge giving a grey appearance, and
there was a slight shift of origin during the import of the
encapsulated PostScript file for editing by GIMP.

The GIF was generated with the following ImageMagick command.
   convert -density 288 -antialias -resize 31% sun95_gridco.eps sun95_gridco.gif