KAPPA: New figure for FRACTION co-ordinates.
This is Figure 13 is the section describing the various co-ordinate
systems of SUN/95.  It contains an extra row showing FRACTION
co-ordinates in addition to the other systems, and a short description
of the system.  The dark shadow was replaced by a grey edge and some
slight shift of origin during the import of the existing encapsulated
PostScript file for editing by GIMP.

The GIF was generated with the following ImageMagick command.
  convert -density 288 -antialias -resize 31% sun95_fraco.eps sun95_fraco.gif

Add the GIF to the list of extra files to copy to sun95.htx.