smurf: start adding map-based despiker (smf_map_spikes)
Fiddled smf_rebinmap1 to report the scale factor for the
input weights that would result in the same average
variance in each pixel as that obtained from the sample
variance. Also added smf_map_spikes for flagging spikes
in the data that are some threshold sigma away from the
current mean value stored in each map pixel. The
variance scaling will be required so that the threshold
is normalized properly.