KAPPA: Propagate fully mask bad pixels for MFITTREND auto method=simple.
Fix bug whereby some bad pixels in the binned array were not
propagated to the output mask NDF.  This could occur when the number
of pixels along the trend axis was not an exact multiple of the number
of bins.  The final elements in the binned array that combined fewer
original data were being overlooked in the mapping of their bad values
to the mask.

This may have been deliberate in my original coding, as the number of
values used to filter data may be quite small, and the significance of
outliers unreliable.  For large samples it may be better to increase
the number of bins from the default 32 or tune to the data to avoid
just a few pixels leading to random masking of edge elements.

For this kps1_mfeb.gdr required the dimensions of the binned array to
be passed, not evaluated internally as DIMS(I)/COMPRS(I) for the ith
dimension, as this could be one element too few.