surf: Fix problems in ADD_DBM reported by Mike Reid
Two calls to SCULIB_GET_FITS_D should have been
SCULIB_GET_FITS_R. This was causing corruption of the PIXSIZE
variable and was resulting in NaN value in the FITS header.

Also noticed that the input variance was being ignored and the
data array was being passed to SURFLIB_CALC_CHOPPED_IMAGE
by mistake.