KAPPA: Do not map signed integer types directly in STATS.
This to enable the ordered statistics to be calculated with KPG_STOSx
that is now restricted to BDIRW instantiations.  This with generic PDA
sorting of these types seemed to be the best approach.  It was not
possible to restore the former behaviour without restoring a fair
chunk of KPG_STOSx to STATS for each type.