smurf: Rewrite smurf internal history checking
SMURF subroutines can modify data and add history tags
to indicate to other subroutines that a certain operation
has been completed. This is required because you can either
use SMURF as a set of discrete commands or you can make
SMURF do everything in one go.

Previously SMURF wrote these hints to NDF HISTORY and
read them back some time later. Recently this has stopped
working because the App name stored in the history was
the global app name and not the subroutien name that
was required. To prevent further interactions with NDF
HISTORY changes I have rewritten the history handling
code to write simple tags to a _CHAR array in the
SMURF extension. This has the added advantage of
not contaminating the main NDF history.

Note that there is a slight API change as I have decided
not to support the contextual string in addition to
the tag.