smurf: Refactor ra/dec and az/el calculation
The calculation of the RA/DEC and AZ/EL columns are the
same except for the key name in the JCMTSTATE structure.
Use this to allow the calculation to occur in a subroutine.

Also add DRA and DDEC as arcseconds offset from base position
and calculate DAZ and DEL in the same manner without erroneously
referring the offsets to the azel of the first sample (since
that is not particular meaningful).