kappa: kappa: use NDF component name as Y axis symbol in LINPLOT
The NDF label value used to be used as the Y axis Symbol, but this could
be long and include spaces (e.g. "Ta corrected antenna temperature").
This caused problems since when CURSOR is used to get positions from the
Plot, the Symbol attribute ends up being used as a column name in the
output catalogue, and so the Symbol attribute needs to have a short value
that includes no spaces. The Label attribute is unchanged (i.e. it is
still set to the NDF label value).