kaplibs: Ensure default CATFRAME parameter does not loose axes
Various kappa commands use parameter CATFRAME to specify the Frame in
which axis values should be stored within an output catalogue. The
default for CATFRAME used to be to search for a SKY Frame within the
catalogues WCS FrameSet, and use the SKY Frame if found (otherwise using
a PIXEL Frame, or the original base Frame). But some WCS FrameSets do not
have a usable Mapping between the SKY Frame and the current Frame. For
instance, if CURSOR is used to get positions from a LINPLOT display
displaying a spectrum extracted from a cube, the WCS will include a SKY
Frame (inherited form the cube), but the Mapping from screen coords to
(RA,Dec) will be degenerate (i.e. all points in the spectrum have the
same (RA,Dec)). To get round this, the default for CATFRAME has been
changed to be the current Frame in the catalogues WCS information.