KAPPA: Report the NDF dimensionality in BLOCK prior to obtaining BOX parameter.
If fewer box sizes are supplied than the number of dimensions of the
NDF, then the final value is duplicated for the missing dimensions.
In case this is not desired, a message at normal level now informs the
user of the NDF dimensionality, if it is more than one, before
prompting for Parameter BOX.  The parameter system already re-prompts
should too many BOX values be supplied.