ndg: Add facilities for recording group contents in NDF history
To use these facilities, create a "GRP NDF history" block using NDG_BEGGH
and NDG_ENDGH. Within this block, call NDG_ADDGH to register one or more
GRP groups. When NDG_ENDGH is called, the contents of the registered groups
(as they existed when NDG_ADDGH was called) will be formatted and appended
to the current history record in each NDF to which default history information
was written during the course of the GRP NDF history block.