CONVERT: Added Group heading to the list of records not indented.
This allows for the new NDG/KAPLIBS facility by which a list of NDFs
stored in a GRP group are listed in a a history block.  The change
means "Group:" is recognised as a heading and thus a new line is
created in the output NDF HISTORY record, indented left of subsequent
lines in the block.  The aim of this and an earlier fix
(3154f4c5771cd) permit moving between FITS and NDF without losing the
block structure and indentation in the HISTORY records.