CONVERT:  Rework the paragrapgh splitting for history records.
Group headings are different from those previously managed, as more
than one can appear in a HISTORY record.  Allow for that and remove
a leading blank to ensure "Group:" is treated as a heading.

Also improved to paragraph parsing to allow for multiple headings
in the buffer portion extracted from a history paragraph.