cupid: Avoid creating an edge of bad pixels in findclumps output mask.
Previously, findclumps created a rim of bad pixels within the output mask
array. This meant that it could avoid certain bounds checking (and so be
a bit faster), since the usual test for bad pixels would detect edge
pixels and so cause the pixel to be avoided. But a rim of bad pixels is
the output mask is bad, particularly when you are processing tiled data
that needs to abut exactly at the edges. So findclumps now does proper
bounds checking rather than creating a rim of bad pixels to detect the
edges. It doesn't seem to have caused any detectable slow-down.