ndg: Document GRP history stuff & allow gh and pv blocks to be halted
Halting a block means that any later NDFs to be accessed are not
recorded, and so will not recieve gh or pv info when the block is ended.
A halted block can be re-started again. This allows applications to exempt
selected NDFs from gh or pv info recording. For instance, I'm about to
use this to prevent provenance and group history being recorded in the
potentially thousands of clump cut-out NDFs created by cupid:findclumps.