smurf: add ability to iterative map-maker to produce single-bolo maps
At the end of solving for all of the model components the iterative
map-maker can now automatically spit out single-detector maps for
every bolo that seemed to be working. It just does a straight rebinning
of the residual signal with the AST model component added back in (i.e.
it should be a clean signal with mostly white noise + astronomical signal;
the other noise components will have been removed in the total map solution
using all of the detectors). Set bolomap=1 in the config file.