smurf: update handing of quality mask bits
The usage of the various SMF__Q_* mask bits was getting very confusing and
inconsistent. Added 3 new macros for collections of mask bits that are
commonly used. SMF__Q_GOOD is for truly good samples that we'd like to
go into the map estimate and calculate chi^2. SMF__Q_MOD are all samples
that we would like to modify as we fit model components. For example,
even if we decide not to use samples taken while the telescope was
stationary (SMF__Q_STAT) in the final map, we will still want to
include such regions when we filter and remove extra noise. Similarly,
we use another special mask SMF__Q_FIT to indicate samples that
should be included whenever we are measuring time-domain model components
(again, we would normally want to include data for such operations
even when the telescope is stationary).