smurf: Check units consistency inside smf_iteratemap
We were checking outside iteratemap but that was not dealing
with flatfielding so you would be unable to combine a flatfielded
data set with a raw data set.

Additionally, the units written to the output map were from the
raw unflatfielded files since smf_open_file was used. Change things
so that smf_iteratemap checks units and returns the units for
writing to the file.