smurf: Fix extinction correction with bad TCS data
Sometimes we have no airmass in a step. Rather than throw out
the data we apply the previous good airmass and continue on. This
is fine in that the data have the correct scaling when it
comes to FFTs of the time series. Won't have an effect on the
map calibration because these data have no telescope position.