KAPPA: State the initial default for interpolation method.
Most uses of the METHOD parameter to describe the interpolation method
uses a default of current value, but didn't state what the initial
default was.

SLIDE docimentation had a fixed default instead of current value.
Also PLUCK's initial default in the interface file was not one of
METHOD's allowed options (although it would be in other tasks).
So set it to SincSinc like the majority of other instances.