smurf: add exportsetbad option to iterative map-maker config file
Setting the new exportsedbad flag will set the appropriate VAL__BAD*
values in model components where bad bolos were identified during
iterative map-making. Currently doesn't do anything with SMF__DKS
or SMF__COM model components (the first because it is trickier
and seldom used, and the latter because it's not clear what is
meant) -- in both of these cases a warning message will be generated
if MSG__VERB is set.