smurf: Remove unused sc2store routines
Remove routines that are not used by the DA or by SMURF. They can
sometimes be confusing when trying to add new features and it is
easier to simply remove them and bring them back later on if we
need to use them.

Most are support routines that were required before SMURF existed.