smurf: add new dimmconfig appropriate for short (1s -- 10s) scans
The default dimmconfig has a high-pass filter in it with a knee
at 0.1 Hz. This won't work for very short scans (i.e. <10s scans
for focus observations). In this config filt the filter is
simply removed. Also note the 100 sample apodization window at
the start and end: this will a reduce a 10s observation to 9s
of usable data in the final map.