smurf: use 'varmapmethod' in bolomap variance generation
Was not previously using the varmapmethod flag when creating
individual bolo maps. Also had to remove a SMF__MINSTATSAMP
check in smf_rebinmap1 in the case that we are propagating
time-stream errors (we can and should estimate a variance
in this case, even with only a single hit in a pixel).