smurf: generate special status for bad frequencies
If smf_get_findex is handed a frequency that will be rounded to 0
Hz, or above the Nyquist frequency, now generate SMF__INFREQ
status that can be trapped later on. These are two potentially
common situations when specifying frequencies for low/high-pass
filters that are inconsistent with the length of the sample
and/or sample rate.