smurf: more flexibility for flagging bad bolos in map-maker
The threshold for the COM model component calculation to reject
bad bolos by checking for outlier gains and/or correlation coefficients
with respect to the common-signal is now configurable: com.corr_tol,
com.gain_tol, com.gain_abstol. Modified the default dimmconfig to use
harsher constraints (going to 3-sigma from 5-sigma) to greatly improve
850um maps. Also cut at 0.2 Hz instead of 0.1 Hz to reduce some more
of the junk. Finally, turn spike detection back on for NOI model component -
it actually seems to work pretty well now for the detectors that are
flagged as "good".