smurf: Do not propagate provenance if there is nothing to propagate
Modify the logic so that we do not propagate unless

 a) we have provenance in the source NDF
or
 b) we have an OBSIDSS to use as a root

This was triggered by STACKFRAMES which can read in 1000 files
that lack provenance information (possibly they should have it!)
but it takes a long time to write out the provenance for 1000
files, especially when it is being written in an incremental
manner.

Should seriously consider modifying smf_updateprov and smf_accumulate_prov
to work with a provenance object and only write it to the output file
at the end.

Also allow the fitshdr to be null.