KAPPA: Fix bug in BEAMFIT interface mode processing planes of a cube.
In interface mode shift the origin of degenerate axes to 1.
Thus it is now possible to fit to arbitrary planes of a cube or
hypercube.  It was previously assuming bounds of 1:1 for the
degenerate axes when taking a section for fitting, and hence
obtaining an image of bad values when the bounds were different.