smurf: add ast.zero_lowhits + updated config files
ast.zero_lowhits will set boundary regions in the map to 0 before
projecting the map into the time domain after each iteration
(part of the AST component). Should help a bit with the zero-point
of the map and systematics of small maps. Also updated
dimmconfig and dimmconfig_faint slightly -- main difference is
FLT is now before AST. Observations of bright sources aren't quite
as good (shadows), but the solution appears to be better behaved in
most other respects.