KAPPA: Move from interim solution to SGP/38 description of <POLYNOMIAL> BSPLINE variant.
For a simple pragmatic solution, I had used XKNOTS and YKNOTS to store
the knots along the X and Y axes.  While this is extensible to a third
dimension, the naming is inelegant and is not generic.  So when I
modified SGP/38 to include the BSPLINE variant of the <POLYNOMIAL>
structure, I used a single vector component called KNOTS that stores a
concatenated list of knots for all axes.  These changes implements
this design change in the code.  Note for backwards compatibility, the
XKNOTS and YKNOTS components continue to be supported when read from
existing data.