CONVERT: Handle periods in ancestor path when no file extension is given.
Data processing at CADC is done such that the path includes at least one
period, and the provenance entries don't have file extensions. In such
cases, the resulting PRVn headers were coming out as strings with eight
spaces in them. This patch checks to see if the last period comes before
the last slash, and if it does, sets the position of the last period to be
the end of the string (which would be where the file extension would be
had there been one).