KAPPA: Use DAT_THERE (as intended) to test for component's existence.
Despite having used it dozens of times, my mind drew a blank trying to
recall the name of this routine when coding at home without the paper
manual. [Only DAT_EXIST came to mind.  SUN/92 could do with a
classified list of routines.] So I put in a temporary workaround to
test the new code.