smurf: add correlated DC step finding/fixing
Optional flag to the DC step finder to fix/flag all bolometers at
locations of steps. This will be useful as there seems to be strong
crosstalk -- the idea is to find all the big steps, and then explicitly
remove small steps in all other detectors at those spots. Also changed
SMF__Q_GOOD so that DC steps will also be ignored in the final map.