smurf: More flexible responsivity calculation
The responsivity can now be calculated using polynomials of order
1 to 3 and can include variance weighting. The polynomial fitting
code has been removed from the responsivity routine to a separate
function.

Can also expand the polynomial to a data cube and now stores
that in an extension of the flatfield output file to provide
direct comparison of the data with the fit.

Currently the error estimates of the responsivity for polynomials
of order greater than 1 seem to be too low by a factor of a few
thousand which messes up the signal-to-noise ratio cut.