smurf: Enable polynomial flatfields
POLYNOMIAL flatfields are now an option in CALCFLAT. This has
required changes to a number of routines to add an additional
argument describing the flatfield type. The responsivity
calculation now correctly determines the gradient at the
reference bolometer value.

A couple of things to note:

 o Anything other than a linear fit has not been tested and
   seems to go wonky for some bolometers.
 o There is no signal-to-noise clipping of responsivities
   when calculating the flatfield in POLYNOMIAL mode. This
   is because the POLYNOMIAL solution does not store any
   error information.

We can get around the second one by ignoring the polynomial
when calculating the responsivity and simply redoing the
fit with errors.

To apply these fits to older data use the COPYFLAT command.