smurf: Correct detection of usable scans in distortion script.
The script used to assume a scan was usable (i.e. had usable WCS) if the
FITS keyword SHUTTER was non-zero.  But some SHUTTER==1 scans do not have
usable WCS and so the test has been changed to "SEQ_TYPE==OBS_TYPE". If
either of these keywords do not exist in the scan header, then the test
reverts to SHUTTER==1.