smurf: renamed SMF__Q_BADS to SMF__Q_BADDA, added SMF__Q_COM
It was getting confusing when data were being flagged by the
common-mode rejection (in smf_calcmodel_com) vs. flagged
by the DA system and/or flatfield. So re-named SMF__Q_BADS
to SMF__Q_BADDA to be more explicit, and used SMF__Q_COM
in smf_calcmodel_com. Also changed a few quality flag
assignments to "|=" from "=" in smf_flatted and
smf_calcmodel_com as this seemed to be causing some odd
behaviour.