First attempt at replacing one_find_file with psx_wordexp globally
The main problem with this patch is that PSX_WORDEXP will return
"*.f" if there are no files matching "*.f" so the caller does
not know that the glob failed. ONE_FIND_FILES would return a
ONE__NOFILES status to indicate no returns from ls.

In most cases the lack of special state is not a problem because
the caller would open the file and fail but testing is required.