smurf: Check for bad COM values in smf_calcmodel_com.
If all bolometers are rejected for a block, the COM values within that
block are set to VAL__BADD. Previously, this was not checked for
resulting in the VAL__BADD values corrupting all the gain and offset
values on the next iteration, which in turn caused all boloemeters
everywhere to be rejected. This commit fixes this problem.