smurf: Use SNR filter for polynomial flatfields
I had forgotten to filter the bad linear fits in POLYNOMIAL
flatfield mode. This meant that some very dody bolometers were
passing the fit and some of them passed the responsivity filter.

Now the linear mode checks the SNR of the gradient.

Note that cubic fits still do not have an SNR check.