ndg: Fix lack of ONE__NOFILES status in ndg1_appen.f
The use of ONE_FIND_FILE within NDG has recently been replaced by
PSX_WORDEXP, but PSX_WORDEXP does not return an error status if no files
are found that match a supplied wild-card template. Instead, it just
retruns a file template that includes the orignal wild-cards. This can be
passed on to the NDF lirary, resulting in errors. To avoid this NDG now
checks to see if the string returned by WORDEXP includes any wild-cards,
and if so, ignores the returned string.