ndg: Improve check for two ancestors being the same.
Previously, two ancestors were considered to be the same if they had the
same path and the same date. This is not bullet-proof since the date is
only accurate to the hearest second, so if an NDF is created, destroyed
and a new NDF with the same path created all within one second, the two
NDFs will be considered to be the same. This patch improves this by also
requiring the CREATOR field to be the same. This is only a partial fix,
since if an application creates two NDFs with the same path within a
space of one second, they will still be considered to be the same. Really
we want some way of associating a globally unique identifier with every
NDF.