smurf: Change matched filter script to use a total PSF sum of unity.
The script used to normalise the psf to a peak value of unity. This is
appropriate for input images in units of "pW/beam", but not for the units
actually created by makemap (pW/bolo). So I've changed the script so
that, by default, it normalises the PSF to a total data sum of unity.
I've added an option ("-np") that gives the old behaviour.