smurf: Add facility to matched-filter script to use a double Gaussian PSF.
As suggested by Ed in an email on 8th April with subject "filtering". The
double Gaussian PSF is formed by creating the normal single Gaussian PSF
with FWHM matching the telescope beam, and then subtracting from it a
Gaussian PSF with a user-specified sigma. Prior to taking their
difference, the primary and secondary Gaussians both have a total data
value of unity. After the difference is taken, the resulting PSF is
re-normalised to have a total data sum of unity.