ndg: Change default AUTORPOV value to NO for foreign format NDFs.
Peter reports problems when using CCDPACK to process many foreign format
NDFs. For Nin input NDFs and Nout output NDFs, a total of Nin*Nout format
conversions need to be performed to update the provenance of all output
NDFs. This can take a long time, so the default for AUTOPROV has been
changed so that the provenance in foreign format output NDFs is only
updated if AUTOPROV is set explicitly to '1'. In addition, provenance
within foreign format input NDFs is ignored unless AUTOPROV is set
explicitly to '1'.

The default behaviour for native NDFs remains unchanged.