smurf: Use template fitting to throw out bad fits in flatfielding
The errors on flatfield data are quite large and with a linear
fit it is possible to fit a curvy flatfield with a high signal-to-noise
straight line. This is problematic so we use smf_templateFit1D
to calculate a correlation coefficient and filter out the
lower end of the distribution.

Note that this only happens for linear fits.