smurf: Constrain sky fit at both ends of flatfield ramp
The first measurement in a flatfield ramp is one step above the reference
heater setting. When the sky variation fit is performed it was not
anchored at the start of the measurement and so was a bit floaty
causing steps in the coadded ramp. We now extrapolate back to the reference
value from the first measurement so that both ends of the fit can
be constrained.