kaplibs: New C function kpg1Config gets a config keymap with name checking and defaults
A text file is used to define all the allowed config parameter and their
default values. A group expression is then obtained from the user containing
config parameter values. An error is reported if the group contains any
unknown config parameters. This allows the defaults to be defined in a
single place, rather than scattered around the application code.